+++ SharePointFramework/ProjectExtensions/src/extensions/projectSetup/index.ts
@@ -9,7 +9,7 @@ import { MessageBarType } from 'office-ui-fabric-react/lib/MessageBar'
 import { ListLogger } from 'pp365-shared/lib/logging'
 import { PortalDataService } from 'pp365-shared/lib/services'
 import * as strings from 'ProjectExtensionsStrings'
-import * as React from 'react'
+import { createElement } from 'react'
 import * as ReactDOM from 'react-dom'
 import { default as HubSiteService } from 'sp-hubsite-service'
 import {
@@ -23,8 +23,10 @@ import {
 } from '../../components'
 import { ListContentConfig, ProjectExtension, ProjectTemplate } from '../../models'
 import * as Tasks from '../../tasks'
+import { deleteCustomizer } from './deleteCustomizer'
 import { ProjectSetupError } from './ProjectSetupError'
 import { IProjectSetupData, IProjectSetupProperties, ProjectSetupValidation } from './types'
+
 export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetupProperties> {
   private _portal: PortalDataService
   private _placeholderIds = {
@@ -52,7 +54,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
       // eslint-disable-next-line default-case
       switch (this._validation) {
         case ProjectSetupValidation.InvalidWebLanguage: {
-          await this._deleteCustomizer(this.componentId, false)
+          await deleteCustomizer(this.context.pageContext.web.absoluteUrl, this.componentId, false)
           throw new ProjectSetupError(
             'onInit',
             strings.InvalidLanguageErrorMessage,
@@ -118,7 +120,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
         iconName: 'Page'
       })
       await this._startProvision(taskParams, data)
-      await this._deleteCustomizer(this.componentId, true)
+      await deleteCustomizer(this.context.pageContext.web.absoluteUrl, this.componentId, true)
     } catch (error) {
       Logger.log({
         message: '(ProjectSetup) [_initializeSetup]: Failed initializing setup',
@@ -137,7 +139,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
   private _getProvisioningInfo(data: IProjectSetupData): Promise<ITemplateSelectDialogState> {
     return new Promise((resolve, reject) => {
       const placeholder = this._getPlaceholder('TemplateSelectDialog')
-      const element = React.createElement<ITemplateSelectDialogProps>(TemplateSelectDialog, {
+      const element = createElement<ITemplateSelectDialogProps>(TemplateSelectDialog, {
         data,
         version: this.manifest.version,
         onSubmit: (state: ITemplateSelectDialogState) => {
@@ -166,7 +168,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
    */
   private _renderProgressDialog(props: IProgressDialogProps) {
     const placeholder = this._getPlaceholder('ProgressDialog')
-    const element = React.createElement<IProgressDialogProps>(ProgressDialog, {
+    const element = createElement<IProgressDialogProps>(ProgressDialog, {
       ...props,
       version: this.manifest.version
     })
@@ -182,7 +184,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
     const progressDialog = this._getPlaceholder('ProgressDialog')
     this._unmount(progressDialog)
     const placeholder = this._getPlaceholder('ErrorDialog')
-    const element = React.createElement<IErrorDialogProps>(ErrorDialog, {
+    const element = createElement<IErrorDialogProps>(ErrorDialog, {
       ...props,
       version: this.manifest.version,
       onDismiss: () => this._unmount(placeholder),
@@ -193,6 +195,8 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
 
   /**
    * Start provision
+   *
+   * Get tasks using Tasks.getTasks and runs through them in sequence
    *
    * @param {Tasks.IBaseTaskParams} taskParams Task params
    * @param {IProjectSetupData} data Data
@@ -237,33 +241,6 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
    */
   private _onTaskStatusUpdated(text: string, subText: string, iconName: string) {
     this._renderProgressDialog({ text, subText, iconName })
-  }
-
-  /**
-   * Delete customizer
-   *
-   * @param {string} componentId Component ID
-   * @param {boolean} reload Reload page after customizer removal
-   */
-  private async _deleteCustomizer(componentId: string, reload: boolean): Promise<void> {
-    const web = new Web(this.context.pageContext.web.absoluteUrl)
-    const customActions = await web.userCustomActions.get<
-      { Id: string; ClientSideComponentId: string }[]
-    >()
-    for (let i = 0; i < customActions.length; i++) {
-      const customAction = customActions[i]
-      if (customAction.ClientSideComponentId === componentId) {
-        Logger.log({
-          message: `(ProjectSetup) [_deleteCustomizer]: Deleting custom action ${customAction.Id}`,
-          level: LogLevel.Info
-        })
-        await web.userCustomActions.getById(customAction.Id).delete()
-        break
-      }
-    }
-    if (reload) {
-      window.location.href = window.location.href
-    }
   }
 
   /**
@@ -391,7 +368,7 @@ export default class ProjectSetup extends BaseApplicationCustomizer<IProjectSetu
    * @param {Web} hubWeb Hub web
    * @param {string} listName List name
    */
-  private _initializeSPListLogging(hubWeb: Web, listName = 'Logg') {
+  private _initializeSPListLogging(hubWeb: Web, listName: string = 'Logg') {
     ListLogger.init(
       hubWeb.lists.getByTitle(listName),
       {