+++ SharePointFramework/ProjectExtensions/src/tasks/SetupProjectInformation/index.ts
@@ -65,7 +65,13 @@ export class SetupProjectInformation extends BaseTask {
   }
 
   /**
-   * Add entry to hub
+   * Add entry to hub project list
+   *
+   * Stores the project with
+   * * Title
+   * * GtSiteId
+   * * GtProjectTemplate
+   * * ContentTypeId (if custom content type is specified in template parameters)
    *
    * @param {IBaseTaskParams} params Task parameters
    */
@@ -78,25 +84,25 @@ export class SetupProjectInformation extends BaseTask {
         params.context.pageContext.legacyPageContext.groupId
       )
       if (entity) return
-      const item: TypedHash<any> = {
+      const properties: TypedHash<any> = {
         Title: params.context.pageContext.web.title,
-        GtSiteId: params.context.pageContext.site.id.toString()
+        GtSiteId: params.context.pageContext.site.id.toString(),
+        GtProjectTemplate: this.data.selectedTemplate.text
       }
       if (params.templateSchema.Parameters.ProjectContentTypeId) {
-        item.ContentTypeId = params.templateSchema.Parameters.ProjectContentTypeId
+        properties.ContentTypeId = params.templateSchema.Parameters.ProjectContentTypeId
       }
       this.logInformation(
         `Adding project entity to list '${params.properties.projectsList}' at ${this.data.hub.url}`,
-        { item }
+        { properties }
       )
       await params.entityService.createNewEntity(
         params.context.pageContext.legacyPageContext.groupId,
         params.context.pageContext.web.absoluteUrl,
-        item
+        properties
       )
       this.logInformation(
-        `Project entity added to list '${params.properties.projectsList}' at ${this.data.hub.url}`,
-        {}
+        `Project entity added to list '${params.properties.projectsList}' at ${this.data.hub.url}`
       )
     } catch (error) {
       throw error