+++ SharePointFramework/ProjectWebParts/src/components/ProjectPhases/ChangePhaseDialog/types.ts
@@ -1,27 +1,19 @@
-import { Phase, IPhaseChecklistItem } from 'models'
-import { IDialogProps } from 'office-ui-fabric-react/lib/Dialog'
+import { IProjectPhaseChecklistItem } from 'pp365-shared/lib/models'
 import { View } from './Views'
 
-export interface IChangePhaseDialogProps extends IDialogProps {
-  /**
-   * New phase
-   */
-  newPhase: Phase
-
-  /**
-   * Active phase
-   */
-  activePhase: Phase
-
-  /**
-   * On change phase
-   */
-  onChangePhase: (phase: Phase) => Promise<void>
-}
-
 export interface IChangePhaseDialogState {
-  isLoading?: boolean
-  checklistItems?: IPhaseChecklistItem[]
+  /**
+   * Check list items
+   */
+  checklistItems?: IProjectPhaseChecklistItem[]
+
+  /**
+   * Currently selected index
+   */
   currentIdx?: number
-  currentView?: View
+
+  /**
+   * Currently selected view
+   */
+  view?: View
 }